Clean up header includes and renamed json_reader to JsonReader to follow filename > classname convention
