+++ src/engine/InputManager.cpp
@@ -20,7 +20,6 @@
 #include <SDL2/SDL_events.h>
 
 #include "App.h"
-#include "InputManager.h"
 
 ohw::InputManager::InputManager() {
 	// Setup the default bindings
