+++ src/engine/Menu.cpp
@@ -17,7 +17,6 @@
 
 #include "App.h"
 #include "engine.h"
-#include "InputManager.h"
 #include "Menu.h"
 #include "Map.h"
 #include "graphics/display.h"
