+++ src/engine/game/actor_pig.cpp
@@ -22,7 +22,6 @@
 #include "player.h"
 #include "actor_manager.h"
 #include "actor_pig.h"
-#include "InputManager.h"
 
 REGISTER_ACTOR( ac_me, APig )    // Ace
 REGISTER_ACTOR( gr_me, APig )    // Grunt
