+++ src/engine/graphics/display.cpp
@@ -16,7 +16,6 @@
  */
 
 #include "engine.h"
-#include "InputManager.h"
 #include "imgui_layer.h"
 #include "Menu.h"
 #include "Map.h"
